Lua v4 (read-only Ausbau) + DumpProto Linux-Fix (#83)
- Lua-API: player.getExp/Stamina/Stats/AttackPower/Defense, map.getName,
  skill.* (read-only Cooldowns, KEIN cast), party.* (eigene Party), quest.*;
  ui.createLine/createPanel/setLayer; F10 per-script Fehler + Einzel-Reload.
  Alle read-only, Sandbox-Kern unangetastet. Security-Review: clean.
- DumpProto (#83): DWORD/long -> uint32/int32 (LP64-Layout-Fix), tea-Casts;
  ermoeglicht Linux-Proto-Rebuild (neue vnums 4101/4201 client-sichtbar).
- 3 neue Beispiel-Addons (char_panel/cooldown_tracker/party_panel) + README.

+++ src/DumpProto/dump_proto.cpp
@@ -28,7 +28,9 @@
 
 typedef unsigned char BYTE;
 typedef unsigned short WORD;
-typedef unsigned long DWORD;
+// See lzo.h: DWORD must be a fixed 4-byte type to match the client's on-disk
+// proto record layout (stride 335 mob / 236 item) and the 16-byte DWORD[4] key.
+typedef uint32_t DWORD;
 
 enum EMisc
 {
@@ -163,7 +165,7 @@ enum EItemMisc
 typedef struct SItemLimit
 {
 	BYTE	bType;
-	long	lValue;
+	int32_t	lValue;		// 4 bytes: matches client (Win32 `long`), NOT LP64 `long`
 } TItemLimit;
 #pragma pack()
 
@@ -171,7 +173,7 @@ typedef struct SItemLimit
 typedef struct SItemApply
 {
 	BYTE	bType;
-	long	lValue;
+	int32_t	lValue;		// 4 bytes: matches client (Win32 `long`), NOT LP64 `long`
 } TItemApply;
 #pragma pack()
 
@@ -198,8 +200,8 @@ typedef struct
 
 	TItemLimit	aLimits[ITEM_LIMIT_MAX_NUM];
 	TItemApply	aApplies[ITEM_APPLY_MAX_NUM];
-	long        alValues[ITEM_VALUES_MAX_NUM];
-	long	alSockets[ITEM_SOCKET_MAX_NUM];
+	int32_t     alValues[ITEM_VALUES_MAX_NUM];	// 4-byte each (Win32 long)
+	int32_t	alSockets[ITEM_SOCKET_MAX_NUM];		// 4-byte each (Win32 long)
 	DWORD	dwRefinedVnum;
 	WORD	wRefineSet;
 	BYTE	bAlterToMagicItemPct;
